requirements.txt
# -----------------------------
# AI & Agent Dependencies
# -----------------------------
crewai==1.9.3
pydantic==2.11.0
openai==1.83.0

# -----------------------------
# Web / Backend
# -----------------------------
fastapi==0.111.1
uvicorn==0.24.0
sqlalchemy==2.0.21
psycopg[binary]==3.2.9
requests==2.32.0
beautifulsoup4==4.12.2
python-dotenv==1.0.1

# -----------------------------
# Containerization & Dev Tools
# -----------------------------
docker==6.1.3
pytest==7.4.2
gitpython==3.1.33

# -----------------------------
# AI Vector / Search Tools
# -----------------------------
faiss-cpu==1.7.4

# -----------------------------
# Optional / Utilities
# -----------------------------
tiktoken==0.4.0  # Only if using GPT tokenization
